# back-end/docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/users/login": {
      "post": {
        "operationId": "UsersController_login",
        "parameters": [
          {
            "name": "x-user-role",
            "in": "header",
            "required": false,
            "description": "Role: SUPER_USER | ADMIN | ORGANIZER | ATTENDEE | STAFF",
            "schema": {}
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "description": "User ID for user-specific operations",
            "schema": {}
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful"
          },
          "400": {
            "description": "Validation error"
          }
        },
        "summary": "Login with email and password",
        "tags": [
          "Users"
        ]
      }
    },
    "/api/users/signup": {
      "post": {
        "operationId": "UsersController_signup",
        "parameters": [
          {
            "name": "x-user-role",
            "in": "header",
            "required": false,
            "description": "Role: SUPER_USER | ADMIN | ORGANIZER | ATTENDEE | STAFF",
            "schema": {}
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "description": "User ID for user-specific operations",
            "schema": {}
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created"
          },
          "409": {
            "description": "Email already exists"
          }
        },
        "summary": "Sign up a new user",
        "tags": [
          "Users"
        ]
      }
    },
    "/api/users": {
      "get": {
        "operationId": "UsersController_findAll",
        "parameters": [
          {
            "name": "x-user-role",
            "in": "header",
            "description": "Role of the requesting user",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "description": "User ID for user-specific operations",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "List of users"
          },
          "403": {
            "description": "Access denied"
          }
        },
        "summary": "List all users",
        "tags": [
          "Users"
        ]
      },
      "post": {
        "operationId": "UsersController_create",
        "parameters": [
          {
            "name": "x-user-role",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "description": "User ID for user-specific operations",
            "schema": {}
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created"
          }
        },
        "summary": "Create a new user (admin)",
        "tags": [
          "Users"
        ]
      }
    },
    "/api/users/{id}": {
      "get": {
        "operationId": "UsersController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "description": "User ID for user-specific operations",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "User details"
          },
          "404": {
            "description": "User not found"
          }
        },
        "summary": "Get user by ID",
        "tags": [
          "Users"
        ]
      },
      "patch": {
        "operationId": "UsersController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "description": "User ID for user-specific operations",
            "schema": {}
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User updated"
          }
        },
        "summary": "Update user",
        "tags": [
          "Users"
        ]
      },
      "delete": {
        "operationId": "UsersController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "description": "User ID for user-specific operations",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "User deleted"
          }
        },
        "summary": "Delete user",
        "tags": [
          "Users"
        ]
      }
    },
...
    "/api/resources": {
      "get": {
        "operationId": "ResourcesController_findAll",
        "parameters": [
          {
            "name": "x-user-role",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "description": "User ID for user-specific operations",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "List of resources"
          }
        },
        "summary": "List all resources",
        "tags": [
          "Resources"
        ]
      },
      "post": {
        "operationId": "ResourcesController_create",
        "parameters": [
          {
            "name": "x-user-role",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "description": "User ID for user-specific operations",
            "schema": {}
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateResourceDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Resource created"
          }
        },
        "summary": "Create a new resource",
        "tags": [
          "Resources"
        ]
      }
    },
    "/api/resources/{id}": {
      "get": {
        "operationId": "ResourcesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "description": "User ID for user-specific operations",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "404": {
            "description": ""
          }
        },
        "summary": "Get resource by ID",
        "tags": [
          "Resources"
        ]
      },
      "patch": {
        "operationId": "ResourcesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "description": "User ID for user-specific operations",
            "schema": {}
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateResourceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Resource updated"
          }
        },
        "summary": "Update resource",
        "tags": [
          "Resources"
        ]
      },
      "delete": {
        "operationId": "ResourcesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "description": "User ID for user-specific operations",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "Resource deleted"
          }
        },
        "summary": "Delete resource",
        "tags": [
          "Resources"
        ]
      }
    },
...
    "/api/registrations/user/{userId}": {
      "get": {
        "operationId": "RegistrationsController_findByUser",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "description": "User ID for user-specific operations",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get registrations for a user",
        "tags": [
          "Registrations"
        ]
      }
    },
...
  },
  "info": {
    "title": "EMCP API",
    "description": "Event Management & Coordination Platform REST API. Role is passed via the x-user-role header. User ID is passed via the x-user-id header for user-specific operations.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "admin@emcp.io"
          },
          "password": {
            "type": "string",
            "example": "password123"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "John Doe"
          },
          "email": {
            "type": "string",
            "example": "john@emcp.io"
          },
          "password": {
            "type": "string",
            "example": "password123"
          },
          "role": {
            "type": "string",
            "example": "ORGANIZER",
            "enum": [
              "SUPER_USER",
              "ADMIN",
              "ORGANIZER",
              "ATTENDEE",
              "STAFF"
            ]
          },
          "status": {
            "type": "string",
            "example": "Active"
          },
          "orgDept": {
            "type": "string",
            "example": "CS Dept"
          }
        },
        "required": [
          "name",
          "email",
          "password",
          "role"
        ]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "orgDept": {
            "type": "string"
          }
        }
      },
...
      "CreateResourceDto": {
        "type": "object",
        "properties": {
          "resourceName": {
            "type": "string",
            "example": "Projector"
          },
          "quantity": {
            "type": "number",
            "example": 10
          }
        },
        "required": [
          "resourceName",
          "quantity"
        ]
      },
      "UpdateResourceDto": {
        "type": "object",
        "properties": {
          "resourceName": {
            "type": "string"
          },
          "quantity": {
            "type": "number"
          }
        }
      },
...
    }
  }
}